+++ source/GenomicsAnalysisCode/resources/omics/import_variant_lambda.py
@@ -11,18 +11,21 @@
 # *  and limitations under the License.                                                                                *
 # *********************************************************************************************************************/
 
-from __future__ import print_function
 from crhelper import CfnResource
 import logging
 import boto3
+from botocore.exceptions import ClientError
 
 logger = logging.getLogger(__name__)
 # Initialise the helper, all inputs are optional, this example shows the defaults
 helper = CfnResource(json_logging=False, log_level='DEBUG', boto_level='CRITICAL')
 
+# Initiate client
 try:
-    codebuild = boto3.client('codebuild')
-    # pass
+    print("Attempt to initiate client")
+    omics_session = boto3.Session()
+    omics_client = omics_session.client('omics')
+    print("Attempt to initiate client complete")
 except Exception as e:
     helper.init_failure(e)
 
@@ -30,79 +33,86 @@ except Exception as e:
 @helper.create
 def create(event, context):
     logger.info("Got Create")
-    start_build_job(event, context)
+    start_import_variants_job(event, context)
 
 
 @helper.update
 def update(event, context):
     logger.info("Got Update")
-    start_build_job(event, context)
+    start_import_variants_job(event, context)
 
 
 @helper.delete
 def delete(event, context):
     logger.info("Got Delete")
-    start_build_job(event, context, action='teardown')
+    return "delete"
     # Delete never returns anything. Should not fail if the underlying resources are already deleted. Desired state.
 
 
 @helper.poll_create
 def poll_create(event, context):
     logger.info("Got Create poll")
-    return check_build_job_status(event, context)
+    return get_variant_import_job_status(event, context)
 
 
 @helper.poll_update
 def poll_update(event, context):
     logger.info("Got Update poll")
-    return check_build_job_status(event, context)
+    return get_variant_import_job_status(event, context)
 
 
 @helper.poll_delete
 def poll_delete(event, context):
     logger.info("Got Delete poll")
-    return check_build_job_status(event, context)
+    return "Poll delete"
 
 
 def handler(event, context):
     helper(event, context)
 
 
-def start_build_job(event, context, action='setup'):
-    response = codebuild.start_build(
-        projectName=event['ResourceProperties']['CodeBuildProjectName'],
-        environmentVariablesOverride=[{
-            'name': 'SOLUTION_ACTION',
-            'value': action,
-            'type': 'PLAINTEXT'
-        }]
-    )
+def start_import_variants_job(event, context):
+    variant_store_name = event['ResourceProperties']['VariantStoreName']
+    role_arn = event['ResourceProperties']['OmicsImportVariantRoleArn']
+    variant_items = [{
+        "source": event['ResourceProperties']['VcfS3Uri']
+    }]
+    try:
+        print("Attempt to start variant import job")
+        response = omics_client.start_variant_import_job(
+            destinationName=variant_store_name,
+            roleArn=role_arn,
+            items=variant_items
+            )
+    except ClientError as e:
+        raise Exception( "boto3 client error : " + e.__str__())
+    except Exception as e:
+       raise Exception( "Unexpected error : " +    e.__str__())
     logger.info(response)
+    helper.Data.update({"VariantImportJobId": response['jobId']})
 
-    helper.Data.update({"JobID": response['build']['id']})
-
-
-def check_build_job_status(event, context):
-    code_build_project_name = event['ResourceProperties']['CodeBuildProjectName']
-
-    if not helper.Data.get("JobID"):
-        raise ValueError("Job ID missing in the polling event.")
-
-    job_id = helper.Data.get("JobID")
-
-    # 'SUCCEEDED' | 'FAILED' | 'FAULT' | 'TIMED_OUT' | 'IN_PROGRESS' | 'STOPPED'
-    response = codebuild.batch_get_builds(ids=[job_id])
-    build_status = response['builds'][0]['buildStatus']
-
-    if build_status == 'IN_PROGRESS':
-        logger.info(build_status)
+def get_variant_import_job_status(event, context):
+    variant_store_name = event['ResourceProperties']['VariantStoreName']
+    variant_import_job_id = helper.Data.get('VariantImportJobId')
+    
+    try:
+        response = omics_client.get_variant_import_job(jobId=variant_import_job_id)
+    except ClientError as e:
+        raise Exception( "boto3 client error : " + e.__str__())
+    except Exception as e:
+       raise Exception( "Unexpected error : " +    e.__str__())
+    status = response['status']
+    
+    if status in ['CREATING', 'UPDATING', 'IN_PROGRESS', 'QUEUED', 'SUBMITTED']:
+        logger.info(status)
         return None
     else:
-        if build_status == 'SUCCEEDED':
-            logger.info(build_status)
+        if status in ['READY', 'COMPLETED', 'ACTIVE', 'COMPLETE']:
+            logger.info(status)
             return True
         else:
-            msg = "Code Build job '{0}' in project '{1}' exited with a build status of '{2}'. Please check the code build job output log for more information." \
-                .format(job_id, code_build_project_name, build_status)
+            msg = f"Variant import job {variant_import_job_id} in\
+                 variant store {variant_store_name} has status {status}, exiting"
             logger.info(msg)
             raise ValueError(msg)
+