+++ source/setup.sh
@@ -20,6 +20,7 @@ clone_and_commit() {
   git init .; git remote add origin ${repo_http_url}
 
   git add *; git commit -m "first commit"; git push --set-upstream origin master
+
 }
 
 wait_for_pipeline() {
@@ -66,8 +67,7 @@ copy_test_data() {
 
   local data_lake_bucket=$(aws cloudformation describe-stacks --stack-name ${pipe_stackname} --query 'Stacks[].Outputs[?OutputKey==`DataLakeBucket`].OutputValue' --output text)
 
-  copy_unpack_zip s3://${artifact_bucket}/${artifact_key_prefix}/variants/onekg-chr22-by_sample/chr22.parquet.zip s3://${data_lake_bucket}/variants/parquet/onekg-chr22-by_sample/
-
+  
   copy_unpack_zip s3://${artifact_bucket}/${artifact_key_prefix}/tcga/tcga-clinical.zip s3://${data_lake_bucket}/
   copy_unpack_zip s3://${artifact_bucket}/${artifact_key_prefix}/tcga/tcga-cnv.zip s3://${data_lake_bucket}/
   copy_unpack_zip s3://${artifact_bucket}/${artifact_key_prefix}/tcga/tcga-expression.zip s3://${data_lake_bucket}/
@@ -75,11 +75,11 @@ copy_test_data() {
   copy_unpack_zip s3://${artifact_bucket}/${artifact_key_prefix}/tcga/tcia-metadata.zip s3://${data_lake_bucket}/
   copy_unpack_zip s3://${artifact_bucket}/${artifact_key_prefix}/tcga/tcga-summary.zip s3://${data_lake_bucket}/
   
-  aws s3 cp s3://${artifact_bucket}/${artifact_key_prefix}/variants/vcf/variants.vcf.bgz s3://${data_lake_bucket}/variants/vcf/variants.vcf.bgz
-  aws s3 cp s3://${artifact_bucket}/${artifact_key_prefix}/variants/vcf/part-00000-a73946db-afb7-49a5-8bbb-621cb57637c2-c000.snappy.parquet s3://${data_lake_bucket}/variants/parquet/vcf/part-00000-a73946db-afb7-49a5-8bbb-621cb57637c2-c000.snappy.parquet
-  aws s3 cp s3://${artifact_bucket}/${artifact_key_prefix}/variants/vcf/part-00001-a73946db-afb7-49a5-8bbb-621cb57637c2-c000.snappy.parquet s3://${data_lake_bucket}/variants/parquet/vcf/part-00001-a73946db-afb7-49a5-8bbb-621cb57637c2-c000.snappy.parquet
-  aws s3 cp s3://${artifact_bucket}/${artifact_key_prefix}/annotation/clinvar/variant_summary.txt.gz s3://${data_lake_bucket}/annotation/clinvar/variant_summary.txt.gz
-  aws s3 cp s3://${artifact_bucket}/${artifact_key_prefix}/annotation/clinvar/part-00000-38061af6-1c74-4605-a37d-ba7b260ad06a-c000.snappy.parquet s3://${data_lake_bucket}/annotation/parquet/clinvar/part-00000-38061af6-1c74-4605-a37d-ba7b260ad06a-c000.snappy.parquet
+  aws s3 cp s3://${artifact_bucket}/${artifact_key_prefix}/annotation/clinvar/clinvar.vcf.gz s3://${data_lake_bucket}/annotation/vcf/clinvar/clinvar.vcf.gz
+  aws s3 cp s3://${artifact_bucket}/${artifact_key_prefix}/variants/vcf/variants.vcf.gz s3://${data_lake_bucket}/variants/vcf/variants.vcf.gz
+  aws s3 cp s3://${artifact_bucket}/${artifact_key_prefix}/variants/1kg/ALL.chr22.shapeit2_integrated_snvindels_v2a_27022019.GRCh38.phased.filtNA.vcf.gz s3://${data_lake_bucket}/variants/1kg/ALL.chr22.shapeit2_integrated_snvindels_v2a_27022019.GRCh38.phased.filtNA.vcf.gz
+  aws s3 cp s3://${artifact_bucket}/${artifact_key_prefix}/references/hg38/Homo_sapiens_assembly38.fasta s3://${data_lake_bucket}/references/hg38/Homo_sapiens_assembly38.fasta
+
 }
 
 setup() {
@@ -112,14 +112,14 @@ setup() {
   # Get pipeline name
   pipeline_name=$(aws cloudformation describe-stack-resource --stack-name ${pipe_stackname} --logical-resource-id CodePipeline --query 'StackResourceDetail.PhysicalResourceId')
 
-  # Wait for pipeline execution using commit id
-  wait_for_pipeline "${pipeline_name}" "${commit_id}"
-
   # Copy Test Data
   copy_test_data "${artifact_bucket}" "${artifact_key_prefix}" "${pipe_stackname}"
 
-  # Run Test
-  "${code_dir}/awscli_test.sh" "${resource_prefix}"
+  # Wait for pipeline execution using commit id
+  wait_for_pipeline "${pipeline_name}" "${commit_id}"
+
+  # Run Crawlers for TCGA data
+  "${code_dir}/run_crawlers.sh" "${resource_prefix}"
 }
 
 project_name=${PROJECT_NAME:-GenomicsAnalysis}